pyproject.toml
[build-system]
requires = ['setuptools>=61.1.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = "wad2023-program"
version = "1.2.0-dev"
description = "Script to retrieve the program for WeAreDevelopers 2023 conference"
readme = "README.md"
authors = [{ name = "Daryl Stark", email = "daryl@dstark.nl" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["conference", "program", "wearedevelopers", "wearedevs"]
dependencies = [
    "requests >= 2.27.0",
    "pydantic >= 1.10.0",
    "beautifulsoup4 >= 4.12.0",
    "python-dateutil >= 2.8.0",
    "rich >= 13.4.0",
    "typer >= 0.9.0",
    "pytz >= 2023.3"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    'pytest >= 7.3.0',
    'pycodestyle == 2.10.0',
    'autopep8 == 2.0.0',
    'pylint == 2.17.0',
    'mypy == 1.3.0',
    'flake8 == 6.0.0',
    'pydocstyle == 6.3.0',
]

[project.urls]
homepage = "https://github.com/DarylStark/WeAreDevelopers_2023_Program"
repository = "https://github.com/DarylStark/WeAreDevelopers_2023_Program"

[project.scripts]
wad23 = "wad2023_program.__main__:main"

[tool.bumpver]
current_version = "1.2.0-dev"
version_pattern = "MAJOR.MINOR.PATCH[-TAG]"
commit_message = "Version {new_version}"
commit = true
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/wad2023_program/__init__.py" = ["{version}"]

[tool.pytest.ini_options]
pythonpath = "./src/"
testpaths = [
    "tests"
]